requirements.txt
psycopg2-binary>=2.9.9
openai>=1.40.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
flask==3.0.0
flask-cors==4.0.0
joblib==1.3.2
python-dotenv>=1.0.0